refactor(plugin-workflow): move bind workflow settings to plugin
