+++ packages/core/client/src/modules/actions/submit/CreateSubmitActionInitializer.tsx
@@ -22,9 +22,6 @@ export const CreateSubmitActionInitializer = (props) => {
       type: 'primary',
       htmlType: 'submit',
     },
-    'x-action-settings': {
-      triggerWorkflows: [],
-    },
   };
   return <ActionInitializerItem {...props} schema={schema} />;
 };
