--- packages/plugins/@nocobase/plugin-workflow/src/client/settings/customizeSubmitToWorkflowActionSettings.tsx
@@ -1,65 +1,0 @@
-/**
- * This file is part of the NocoBase (R) project.
- * Copyright (c) 2020-2024 NocoBase Co., Ltd.
- * Authors: NocoBase Team.
- *
- * This project is dual-licensed under AGPL-3.0 and NocoBase Commercial License.
- * For more information, please refer to: https://www.nocobase.com/agreement.
- */
-
-import { useFieldSchema } from '@formily/react';
-import { isValid } from '@formily/shared';
-import {
-  AfterSuccess,
-  AssignedFieldValues,
-  ButtonEditor,
-  RemoveButton,
-  SchemaSettings,
-  SecondConFirm,
-  SkipValidation,
-  WorkflowConfig,
-  useSchemaToolbar,
-} from '@nocobase/client';
-
-export const customizeSubmitToWorkflowActionSettings = new SchemaSettings({
-  name: 'actionSettings:submitToWorkflow',
-  items: [
-    {
-      name: 'editButton',
-      Component: ButtonEditor,
-      useComponentProps() {
-        const { buttonEditorProps } = useSchemaToolbar();
-        return buttonEditorProps;
-      },
-    },
-    {
-      name: 'secondConfirmation',
-      Component: SecondConFirm,
-    },
-    {
-      name: 'assignFieldValues',
-      Component: AssignedFieldValues,
-    },
-    {
-      name: 'skipRequiredValidation',
-      Component: SkipValidation,
-    },
-    {
-      name: 'afterSuccessfulSubmission',
-      Component: AfterSuccess,
-    },
-    {
-      name: 'bindWorkflow',
-      Component: WorkflowConfig,
-    },
-    {
-      name: 'delete',
-      sort: 100,
-      Component: RemoveButton as any,
-      useComponentProps() {
-        const { removeButtonProps } = useSchemaToolbar();
-        return removeButtonProps;
-      },
-    },
-  ],
-});
